node-editor: Use nested GtkBox'es instead of a GtkGrid
GtkGrid shines when we need cells in different rows to have the same
width, and/or cells in different columns to have the same height. We
don't need that here at all, so a box (with another nested box for the
two controls we want to place next to each other horizontally) is more
appropriate. As things are, GtkBox also handles height-for-width layout
much better than GtkGrid.

Fixes https://gitlab.gnome.org/GNOME/gtk/-/issues/7402, sort of, though
the real fix would be in GtkGrid's layout implementation.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>