node-editor: Draw the sidebar images at node size
Previously, they were scaled down to the size they would appear in
in the sidebar. While this can sometimes produce interesting bugs due
to the added scale, it also often hides bugs due to that scale.

So don't do that anymore and make the renderer render a texture in
actual size and then downscale the texture into the sidebar.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8506>