svgpaintable: Use a cairo node
We can convert the svg to a cairo node once, and then we don't
need to keep the RsvgHandle around. Since we use a recording
surface for cairo nodes, the result is still resizable.

This makes the svgpaintable code much more similar to the
svgsymbolicpaintable.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8578>