path paintable: Rename the blur transition
It has been pointed out that this does not look like
a blur, so the name could be confusing. Call it 'morph'
instead, since the effect morphs the first path into a
blob, and then morphs the blob back into the second path.

This is still not ideal naming, since 'morphing' is often
understood to involve path interpolation, which isn't
happening here.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8957>