svg: Use the path parser api
Add a small SvgPathData that is used for interpolating
paths according to the rules of SVG.

The updated test reveals that we are losing a small
amount of efficiency here - stroke bounds for rects
can no longer take advantage of the contour information
and thus have to take miter limits into account.
This could be avoided be recreating the pattern matching
that the path parser does in the svg->gsk path conversion.

Affected tests have been updated.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9299>